requirements.txt
django==3.1.14
djangorestframework==3.13.1
django-extensions==3.1.5
django-timescaledb==0.2.11
psycopg2==2.9.3